@@ -17,7 +17,8 @@
 #ifndef HEADER_SUPERTUX_EDITOR_OBJECT_OPTION_HPP
 #define HEADER_SUPERTUX_EDITOR_OBJECT_OPTION_HPP
 
-#include <boost/optional.hpp>
+#include <optional>
+#include <functional>
 #include <string>
 #include <vector>
 
@@ -75,7 +76,7 @@ class BoolObjectOption : public ObjectOption
 {
 public:
   BoolObjectOption(const std::string& text, bool* pointer, const std::string& key,
-                   boost::optional<bool> default_value,
+                   std::optional<bool> default_value,
                    unsigned int flags);
 
   virtual void save(Writer& write) const override;
@@ -84,7 +85,7 @@ public:
 
 private:
   bool* const m_pointer;
-  const boost::optional<bool> m_default_value;
+  const std::optional<bool> m_default_value;
 
 private:
   BoolObjectOption(const BoolObjectOption&) = delete;
@@ -95,7 +96,7 @@ class IntObjectOption : public ObjectOption
 {
 public:
   IntObjectOption(const std::string& text, int* pointer, const std::string& key,
-                  boost::optional<int> default_value,
+                  std::optional<int> default_value,
                   unsigned int flags);
 
   virtual void save(Writer& write) const override;
@@ -104,7 +105,7 @@ public:
 
 private:
   int* const m_pointer;
-  const boost::optional<int> m_default_value;
+  const std::optional<int> m_default_value;
 
 private:
   IntObjectOption(const IntObjectOption&) = delete;
@@ -150,7 +151,7 @@ class FloatObjectOption : public ObjectOption
 {
 public:
   FloatObjectOption(const std::string& text, float* pointer, const std::string& key,
-                    boost::optional<float> default_value,
+                    std::optional<float> default_value,
                     unsigned int flags);
 
   virtual void save(Writer& write) const override;
@@ -159,7 +160,7 @@ public:
 
 private:
   float* const m_pointer;
-  const boost::optional<float> m_default_value;
+  const std::optional<float> m_default_value;
 
 private:
   FloatObjectOption(const FloatObjectOption&) = delete;
@@ -170,7 +171,7 @@ class StringObjectOption : public ObjectOption
 {
 public:
   StringObjectOption(const std::string& text, std::string* pointer, const std::string& key,
-                     boost::optional<std::string> default_value,
+                     std::optional<std::string> default_value,
                      unsigned int flags);
 
   virtual void save(Writer& write) const override;
@@ -179,7 +180,7 @@ public:
 
 private:
   std::string* const m_pointer;
-  boost::optional<std::string> m_default_value;
+  std::optional<std::string> m_default_value;
 
 private:
   StringObjectOption(const StringObjectOption&) = delete;
@@ -190,7 +191,7 @@ class StringSelectObjectOption : public ObjectOption
 {
 public:
   StringSelectObjectOption(const std::string& text, int* pointer, const std::vector<std::string>& select,
-                           boost::optional<int> default_value,
+                           std::optional<int> default_value,
                            const std::string& key, unsigned int flags);
 
   virtual void save(Writer& write) const override;
@@ -200,7 +201,7 @@ public:
 private:
   int* const m_pointer;
   const std::vector<std::string> m_select;
-  const boost::optional<int> m_default_value;
+  const std::optional<int> m_default_value;
 
 private:
   StringSelectObjectOption(const StringSelectObjectOption&) = delete;
@@ -213,7 +214,7 @@ public:
   EnumObjectOption(const std::string& text, int* pointer,
                    const std::vector<std::string>& labels,
                    const std::vector<std::string>& symbols,
-                   boost::optional<int> default_value,
+                   std::optional<int> default_value,
                    const std::string& key, unsigned int flags);
 
   virtual void save(Writer& write) const override;
@@ -224,7 +225,7 @@ private:
   int* const m_pointer;
   const std::vector<std::string> m_labels;
   const std::vector<std::string> m_symbols;
-  const boost::optional<int> m_default_value;
+  const std::optional<int> m_default_value;
 
 private:
   EnumObjectOption(const EnumObjectOption&) = delete;
@@ -253,7 +254,7 @@ class FileObjectOption : public ObjectOption
 {
 public:
   FileObjectOption(const std::string& text, std::string* pointer,
-                   boost::optional<std::string> default_value,
+                   std::optional<std::string> default_value,
                    const std::string& key,
                    std::vector<std::string> filter,
                    const std::string& basedir,
@@ -266,7 +267,7 @@ public:
 
 private:
   std::string* const m_pointer;
-  boost::optional<std::string> m_default_value;
+  std::optional<std::string> m_default_value;
   const std::vector<std::string> m_filter;
   std::string m_basedir;
   bool m_path_relative_to_basedir;
@@ -280,7 +281,7 @@ class ColorObjectOption : public ObjectOption
 {
 public:
   ColorObjectOption(const std::string& text, Color* pointer, const std::string& key,
-                    boost::optional<Color> default_value, bool use_alpha,
+                    std::optional<Color> default_value, bool use_alpha,
                     unsigned int flags);
 
   virtual void save(Writer& write) const override;
@@ -289,7 +290,7 @@ public:
 
 private:
   Color* const m_pointer;
-  const boost::optional<Color> m_default_value;
+  const std::optional<Color> m_default_value;
   bool m_use_alpha;
 
 private:
