@@ -16,7 +16,7 @@
 
 #include "object/path_object.hpp"
 
-#include <boost/optional.hpp>
+#include <optional>
 
 #include "editor/editor.hpp"
 #include "object/path_gameobject.hpp"
@@ -53,7 +53,7 @@ PathObject::init_path(const ReaderMapping& mapping, bool running_default)
   }
 
   std::string path_ref;
-  boost::optional<ReaderMapping> path_mapping;
+  std::optional<ReaderMapping> path_mapping;
   if (mapping.get("path", path_mapping))
   {
     auto& path_gameobject = d_gameobject_manager->add<PathGameObject>(*path_mapping, true);
