@@ -17,7 +17,7 @@
 #include "video/video_system.hpp"
 
 #include <assert.h>
-#include <boost/optional.hpp>
+#include <optional>
 #include <config.h>
 #include <iomanip>
 #include <physfs.h>
@@ -177,7 +177,7 @@ VideoSystem::do_take_screenshot()
     }
   }
 
-  auto find_filename = [&]() -> boost::optional<std::string>
+  auto find_filename = [&]() -> std::optional<std::string>
     {
       for (int num = 0; num < 1000000; ++num)
       {
@@ -188,7 +188,7 @@ VideoSystem::do_take_screenshot()
           return screenshot_filename;
         }
       }
-      return boost::none;
+      return std::nullopt;
     };
 
   auto filename = find_filename();
